Refactor of the enable/disable system and removal of good practices

+++ Forms/delBp.php
@@ -2,11 +2,23 @@
   session_start(); 
   include '/home/r2c/R2C/Forms/checkSession.php';
 
-  $bpToDelete = $_POST['idbp'];
+  if(isset($_POST['idbp'])) {
+    $bpToDelete = $_POST['idbp'];
 
-  $sql = "DELETE FROM BONNESPRATIQUES WHERE idbp = $bpToDelete";
-  $request = $BDD->prepare($sql);
-  $request->execute();
+    $sql = "DELETE FROM BONNESPRATIQUES_PROGRAMME WHERE idbp = $bpToDelete";
+    $request = $BDD->prepare($sql);
+    $request->execute();
 
-  header('Location: ../Accueil/accueil.php')
+    $sql = "DELETE FROM BONNESPRATIQUES_MOTSCLEF WHERE idbp = $bpToDelete";
+    $request = $BDD->prepare($sql);
+    $request->execute();
+
+    $sql = "DELETE FROM BONNESPRATIQUES WHERE idbp = $bpToDelete";
+    $request = $BDD->prepare($sql);
+    $request->execute();
+
+    header('Location: ../Accueil/accueil.php');
+  } else {
+    header('Location: ../Validation/validation.php?message=ed');
+  }
 ?>
